remove Electron.js completely (plus minor fixes)
