+++ .eslintrc.cjs
@@ -5,8 +5,6 @@ module.exports = {
   extends: [
     'eslint:recommended',
     'plugin:svelte/recommended',
-    '@electron-toolkit/eslint-config-ts/recommended',
-    '@electron-toolkit/eslint-config-prettier',
     'plugin:storybook/recommended'
   ],
   overrides: [
@@ -19,6 +17,5 @@ module.exports = {
     }
   ],
   rules: {
-    'svelte/no-unused-svelte-ignore': 'off'
   }
 }
