+++ .gitignore
@@ -1,6 +1,7 @@
 node_modules
 
 # Output
+out
 .output
 .vercel
 /.svelte-kit
