+++ .prettierrc
@@ -1,5 +1,6 @@
 {
   "tabWidth": 2,
+  "semi": false,
   "singleQuote": true,
   "trailingComma": "es5",
   "printWidth": 100,
