--- .prettierrc.yaml
@@ -1,10 +1,0 @@
-singleQuote: true
-semi: false
-printWidth: 100
-trailingComma: none
-plugins:
-  - prettier-plugin-svelte
-overrides:
-  - files: '*.svelte'
-    options:
-      parser: svelte
