+++ package.json
@@ -16,35 +16,21 @@
     "js:test": "vitest",
     "svelte:check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
     "svelte:check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-  "svelte:dev": "vite dev",
+    "svelte:dev": "vite dev",
     "svelte:build": "vite build",
     "svelte:preview": "vite preview",
     "storybook": "pnpx run storybook:dev",
     "storybook:dev": "storybook dev -p 6006",
-    "storybook:build": "storybook build",
-    "electron:start": "electron-vite preview",
-    "electron:dev": "electron-vite dev",
-    "electron:build": "pnpm run typecheck && electron-vite build",
-    "electron:postinstall": "electron-builder install-app-deps",
-    "electron:build:unpack": "pnpm run build && electron-builder --dir",
-    "electron:build:win": "pnpm run build && electron-builder --win",
-    "electron:build:mac": "pnpm run build && electron-builder --mac",
-    "electron:build:linux": "pnpm run build && electron-builder --linux"
+    "storybook:build": "storybook build"
   },
   "dependencies": {
-    "@electron-toolkit/preload": "^3.0.0",
-    "@electron-toolkit/utils": "^3.0.0",
     "bulma": "^1.0.2",
-    "electron-updater": "^6.1.7",
     "onscan.js": "^1.5.2",
     "open-library-api": "^0.1.2",
     "openapi-fetch": "^0.12.2"
   },
   "devDependencies": {
     "@chromatic-com/storybook": "^2.0.2",
-    "@electron-toolkit/eslint-config-prettier": "^2.0.0",
-    "@electron-toolkit/eslint-config-ts": "^2.0.0",
-    "@electron-toolkit/tsconfig": "^1.0.1",
     "@storybook/addon-essentials": "8.3.3",
     "@storybook/addon-interactions": "8.3.3",
     "@storybook/addon-links": "8.3.3",
@@ -58,9 +44,7 @@
     "@sveltejs/kit": "^2.0.0",
     "@sveltejs/vite-plugin-svelte": "^4.0.0-next.7",
     "@types/eslint": "^9.6.0",
-    "electron": "^32.1.2",
-    "electron-builder": "^25.0.5",
-    "electron-vite": "^2.0.0",
+    "@types/onscan.js": "^1.5.6",
     "eslint": "^9.0.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-config-ts": "^4.0.1",
