--- src-ui/env.d.ts
@@ -1,2 +1,0 @@
-/// <reference types="svelte" />
-/// <reference types="vite/client" />
