--- src-ui/main.ts
@@ -1,11 +1,0 @@
-import('./assets/main.css')
-
-import App from './App.svelte'
-import { mount } from 'svelte';
-
-
-const app = mount(App, {
-  target: document.getElementById('app') as HTMLElement
-})
-
-export default app
