+++ vitest.config.js
@@ -1,8 +1,8 @@
 import { defineConfig } from 'vitest/config'
-import { svelte } from '@sveltejs/vite-plugin-svelte'
+import { sveltekit } from '@sveltejs/kit/vite';
 
 export default defineConfig({
-  plugins: [svelte()],
+  plugins: [sveltekit()],
   test: {
     diff: './vitest.diff.ts'
   }
